Collada (#1)
* Collada loader initial

* Collada OpenGL viewer continue implementation, for model rendering using GLSL shaders

* Collada viewer continue implementation, model added

* Collada viewer implementation continue

* Collada 3D View implemented with test code

* Move static mash from test code in view to shader model

* Texture object added

* Adding precompiled header

* IDE project update

* DirectDraw image loader for textures, initial

* Fixed image loading over DirectX use in memory proxy IStream

* Textured mesh tested

* Normal and Phong shading fixes

* Add image loading from file

* Added normal mapped mesh renderer

* Add RGB texture loading, in addition to RGBA textures

* Texture from image functions

* nm texture change

* Modify normal mapping shaders, to decrease matrix operations in fragment shader

* Remove deffect from diffuse test texture

* Material-collor experiment

* Simplify matrix operations in bumpmapping vertex shader

* Tested on linux, with no texture mesh since no image loader

* Fix Windows build after Unix build changes

* View matrix rotations optimization, remove lookAt etc

* Refactoring

* Start implementing reading textures using GNU libs, for POSIX

* Texture loading over the libpng for Unix

* Image class refactoring

* Rename mesh to surface, add method for reading shaders from file

* Several optimizations after gcov tests

* OpenMP experiment for SIMD

* Fix compile time warning

* Modify file class

* shader.cpp refactoring

* Fix MS VC++ for files.cpp, add missing include header

* Compiler optimization options, gcc

* Several optimizations after gcov tests

* Bug fix, in string blank method implementation

* Reintegrate master

* Fix merge issues

* Refactoring, remove static meshes implementations from model module

* buffer lngrow modification

* Fix files.hpp on UNIX build

* Implement texture mimmap generation with glTexStorage2D

* Adding BGRA pixel format for texture, for future work. (BGRAnot supported by PNG, but used by hardware)

* Change MS VC++ build script to use DLL CRT and VC runtimes, instead of statical linking to those libs

* Refactor shaders from CPP sources

* Refactor shaders and surfaces

* Change OpenGL loader to glad for Linux

* Update GLSL to OpenGL 330, and add GL ES specific guards

* Static mesh to common file shaders

* Fix light positioning in shaders

* Shader update

* Add ability to move light in view

* Perspective to scene and view

* Add perspective correction into sceene

* Add dynamic fov_y

* Simplify fov_y calc (prev was a shame). Correct nm vertex shader to cal eye position correctly

* Blur demo texture

* Make tangetn vector more generic, add ability to calc it with OpenMP

* Fix TODO to GL program validation

* Shader loadeing for file changes. No stack tmp buffer

* Rename static_mesh to mesh for project

* Remove useless isize_ from mesh classes

* Fix for renamed header name

* Bugfix in libpng texture loader, stop on EOF

* Modifications in channels

* Continue with Collada base parser implementation

* Refactor loop in XML reader read_chars

* Texture

* Parser POC code, not working yet

* Phong effect parsing from Collada file

* Modify IDE files

* Continue implementing Collada parser, refactor of pong parsing, add a frame to the root document parsing

* Bug fix, basic Collada frame

* Read of effect_library implemented

* Extend effect loading

* Make effect with the union value

* Change effect unions

* Modify mesh with the assimp like structure. Add  input tag and vertex parsing methods'

* New tag end method added, parser function changed. Source tag parsing implemented. COLLADA model moved to the separate file.

* 1 .cached_string and string pool memory allocator to private windows heap.
2. Collada source to class

* Bug fix in parse_string_array

* Update IDE release build options

* Modify Collada model to classes from raw structs, where appropriate

* Add index data parsing (trangles for now)

* intrusive_array template added

* Parse library_visual_scenes

* 0. Refactor read entity method for reduce calling next
1. Refactor meshes and shader program to bind VBO at once with the vertex array layout

* Add material to material_library e.g. mat-effect-link parsing

* Start loading model from parsed collada structures

* Loading collada parsed structures to OpenGL, continue implementation

* Fix issues with const string comparator

* Intel compiler specific intrinsic and other small fixes after asm code review

* Compilers options refactoring, for asm nano optimization

* String pool bug fix

* Improve loop for entity reading

* Change read characters method, to use source->read_until_char

* cleanup strings.hpp

* Add read until double charchar into xml source, refactor read until double separator using this method.
Test on GCC 8.3.0 Windows

* Bugfix for empty value XML attribute, optimize attribute value normalization

* Refactor parse_sart_element

* Bugfix for skip comment

* Small string optimization (SSO) for const_string class

* Assentation for empty sting with non zero length

* Small string optimization, bug fixes

* Update IDE files, return profiling options into debug

* Use SSO in string pooling

* Optimizations for buffers, for less realloc calls

* Optimizations for read_until_char, after profiling

* Optimize byte_buffer after profiling

* Byte buffer refactoring

* Testing base xml parsing with GCC 9.1.0, fix warnings

* SSO bug fix for 0-ro terminating character

* Change growing strategy for read until char

* Continue Collada parsing implementation

* Add parsing of library_images. 2. Code cleanup

* Read until double separator, pack_word for le/be

* Change refference counting strategry to  acquire-release semantic, for io::object and io::const_string non SSO long buffer

* MS VC++ bugfixes

* Load collada into video memory, first raw implementation

* COLLADA geometry textured_qube.dae loaded to graphical engine

* Extract texture name from COLLADA

* COLLADA newparam tag parsing added

* Add static mesh without color to engine

* Add ability to load multiple sub-meshes

* Start refactoring from experimental code

* Win file, add resolving of a full file path

* Small refactorings

* Add texture loading from COLLADA model file directory

* File path() UNIX implementation

* Refactor model_loader

* Move test models into sub-directory

* Change view movement, add Mouse button + mouse motion to rotate model

* Add bump-mapping meshes loading from COLLADA file

* Update norma map for earth, for more geoid look

* Add open file dialog for open COLLADA file (Windows Ony, Unix uses command line param). WARN gpu folder with shaders needs to be copied into executable directory, i.e. bin/(debug|release)-(os)(arch)

* Add tess_builder

* Refactoring and MS VC++ fixes, including MS VC++ Cmake build

* Cmake VC++ and warning fixes

* Update models files. Fix bug in static model loader

* Fix UNIX build

* Code cleanups

* CppCheck fixes

* Update IDE compiler flags

* Comments

* NURB with tesselation added, not tested yet

* NURB implemented

* polymesh start implementation

* Comments

* Buf fix in view rotation by mouse without left mouse button clicked. And more comments

* Add polygonal textured cube model for polymesh

* Polymesh - adding shaders, not yet tested

* Test for tesselation polygonal mesh (issues with normal vector)

* Add ability to change light type to directional or positional, by pressing space key.

* Work on README.md

* Refactoring

* Refactoring

* String ops refactoring for constexpr where possible, and clenup unused functions

* Fix MS VC++ build

* Fix MS VC++ build for C++ 11 standard, instead of C++17

* Add manual validation functions into cursor (reader) API

* Fix MS VC++ build

* Refactor to use constexpr ctype

* Refactoring, and cleanup

+++ src/win/sockets.cpp
@@ -200,7 +200,7 @@ public:
         const ::addrinfo *ai = static_cast<const ::addrinfo *>(ep_.native());
         if(
             SOCKET_ERROR == ::WSAConnect(s,
-                                         ai->ai_addr, ai->ai_addrlen,
+                                         ai->ai_addr, static_cast<int>(ai->ai_addrlen),
                                          nullptr,nullptr,nullptr,nullptr
                                         )
         ) {
